Correct mistakes from last four commits
This partially reverts fadb680, 8b5b46c and 10b617b.

Corrections @GeryonTom

1. Removed allMissionObjects exception for FunctionsManager. The
exception you added was from 1.8.8. it was changed to entities for 1.8.9
here:

https://github.com/DayZMod/DayZ/commit/29285058e59d9c4a898f4c079aa37c238552df8f#diff-c3c70705c54009eff517eb4723a994dcL53

2. Removed _menu private var list exception. This was another from 1.8.8
which was removed for 1.8.9 in:

https://github.com/DayZMod/DayZ/commit/1d639b557b1b2a57399c113184e17992d678ef7b

3. Removed playableUnits exception which was from 1.8.8. Formatting was
changed for 1.8.9 here:

https://github.com/EpochModTeam/DayZ-Epoch/commit/c641bc20a8fefc1bf09ad9d87214fd11ea6300da#diff-ca5a532a5a88f637bc3eccc5d0ad4edfR24

4. Consolidated three titleCut exceptions to one.

5. Removed createDialog exception that was also for 1.8.8. Updated to
single quote for 1.8.9 here:

https://github.com/DayZMod/DayZ/commit/3a29242537439b866388aeeb63574692d6e9677e

Make sure whoever was kicked was actually updated to 1.8.9 before adding
exceptions for them.

Also when adding exceptions the slash must come before the double quote
to escape it if it is meant to be literal \" In the playableUnits
exception you had it after the " which may have errored out any filters
below that point.